+++ projet_robot/projet_robot/Simulation/Senseur.py
@@ -11,8 +11,8 @@ class Senseur:
         """Renvoie la distance entre le robot et l'obstacle si il est dans le champ de vision du senseur"""
         po=self.portee
         while po >= 0:
-            if ((int)(Robot.x+(po*math.cos(Robot.angle))) >= x and (int)(Robot.y-(po*math.sin(Robot.angle))) >= y) and ((int)(Robot.x+(po*math.cos(Robot.angle))) <= (x+t_x) and (int)(Robot.y-(po*math.sin(Robot.angle))) <= (y+t_y)):
-                return  (int)(math.sqrt(((x-Robot.x)**2)+((y-Robot.y)**2))*0.026)
+            if Robot.x+po*math.cos(Robot.angle) >= x and Robot.y-po*math.sin(Robot.angle) >= y and Robot.x+po*math.cos(Robot.angle) <= x+t_x and Robot.y-po*math.sin(Robot.angle) <= y+t_y:
+                return  math.sqrt((x-Robot.x)**2+(y-Robot.y)**2)*0.026
             po -= 1
         return False
